before finish

+++ app/controllers/MembersController.java
@@ -20,7 +20,8 @@ public class MembersController extends Controller {
     private FormFactory formFactory;
 
     public Result table(Long aid, Long oid){
-        if("" != session("user_id")){
+        boolean role = LoginController.roleAssetAdmin();
+        if("" != session("user_id") && role){
             OilField oilField = OilField.find.where().eq("id", oid).findUnique();
             List<OilUsers> list = OilUsers.find.where().eq("oilField",oilField).findList();
             return ok(Json.toJson(list));
@@ -30,7 +31,8 @@ public class MembersController extends Controller {
     }
 
     public Result add(Long aid,Long oid){
-        if("" != session("user_id")){
+        boolean role = LoginController.roleAssetAdmin();
+        if("" != session("user_id") && role){
             Form<OilUsers> formUser = formFactory.form(OilUsers.class).bindFromRequest();
             OilField oilField = OilField.find.where().eq("id",oid).findUnique();
             OilUsers newUser = formUser.get();
@@ -43,7 +45,8 @@ public class MembersController extends Controller {
     }
 
     public Result edit(Long aid,Long oid,Long mid){
-        if("" != session("user_id")){
+        boolean role = LoginController.roleAssetAdmin();
+        if("" != session("user_id") && role){
             Form<OilUsers> formScen = formFactory.form(OilUsers.class).bindFromRequest();
             OilUsers oilUsers = OilUsers.find.where().eq("id",mid).findUnique();
             oilUsers.login = formScen.get().login;
@@ -57,7 +60,8 @@ public class MembersController extends Controller {
     }
 
     public Result remove(Long aid,Long oid,Long mid){
-        if("" != session("user_id")){
+        boolean role = LoginController.roleAssetAdmin();
+        if("" != session("user_id") && role){
             OilUsers oilUsers = OilUsers.find.where().eq("id",mid).findUnique();
             oilUsers.delete();
             return ok(Json.toJson("Ok"));
@@ -67,7 +71,8 @@ public class MembersController extends Controller {
     }
 
     public Result get(Long aid,Long oid,Long  mid){
-        if("" != session("user_id")){
+        boolean role = LoginController.roleAssetAdmin();
+        if("" != session("user_id") && role){
             OilUsers oilUsers = OilUsers.find.where().eq("id",mid).findUnique();
             return ok(Json.toJson(oilUsers));
         }else{
